requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.8.9"
bcrypt==4.0.0
cffi==1.15.1
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
